+++ src/process_table.c
@@ -6,12 +6,12 @@
 void process_table_init(struct process_table *pt, int hashsize)
 {
     pt->hashsize = hashsize;
-    pt->table = (struct list **)malloc(sizeof(struct list *) * pt->hashsize);
+    pt->table = (struct list **)malloc(sizeof(struct list *) * (size_t)pt->hashsize);
     if (pt->table == NULL)
     {
         exit(1);
     }
-    memset(pt->table, 0, sizeof(struct list *) * pt->hashsize);
+    memset(pt->table, 0, sizeof(struct list *) * (size_t)pt->hashsize);
 }
 
 static int proc_hash(const struct process_table *pt, const struct process *p)
